Fix des coordonnées
Aujourd'hui, j'ai appris que convertir des doubles en longs leur font perdre la virgule.

Mais ça n'explique toujours pas comment mon code marche par contre.

+++ app/src/main/java/com/example/flightstatsm2/Utils.kt
@@ -132,8 +132,8 @@ class Utils private constructor() {
                 val waypointJsonArray: JSONArray = waypointFromObject.getJSONArray(i)
                 val waypointModel = WaypointModel(
                     waypointJsonArray.getInt(0),
-                    waypointJsonArray.getLong(1),
-                    waypointJsonArray.getLong(2),
+                    waypointJsonArray.getDouble(1),
+                    waypointJsonArray.getDouble(2),
                     waypointJsonArray.getLong(3),
                     waypointJsonArray.getLong(4),
                     waypointJsonArray.getBoolean(5)
