+++ app/src/main/java/com/example/flightstatsm2/WaypointModel.kt
@@ -2,8 +2,8 @@ package com.example.flightstatsm2
 
 data class WaypointModel(
     val time: Int,
-    val lat: Long,
-    val long: Long,
+    val lat: Double,
+    val long: Double,
     val altitude: Long,
     val rotation: Long,
     val isOnGround: Boolean
